+++ src/components/Canvas/CanvasInner.default.tsx
@@ -2,6 +2,7 @@ import styled from 'styled-components'
 import { IConfig, IOnCanvasClick } from '../../'
 
 export interface ICanvasInnerDefaultProps {
+  className?: string
   config: IConfig
   children: any
   onClick: IOnCanvasClick