+++ src/components/Link/Link.default.tsx
@@ -1,7 +1,9 @@
 import * as React from 'react'
 import { generateCurvePath, generateRightAnglePath, generateSmartPath, IConfig, ILink, IOnLinkClick, IOnLinkMouseEnter, IOnLinkMouseLeave, IPort, IPosition } from '../../'
+import { ArrowLink, RegularLink } from './variants'
 
 export interface ILinkDefaultProps {
+  className?: string
   config: IConfig
   link: ILink
   startPos: IPosition
@@ -16,66 +18,30 @@ export interface ILinkDefaultProps {
   matrix?: number[][]
 }
 
-export const LinkDefault = ({
-  config,
-  link,
-  startPos,
-  endPos,
-  fromPort,
-  toPort,
-  onLinkMouseEnter,
-  onLinkMouseLeave,
-  onLinkClick,
-  isHovered,
-  isSelected,
-  matrix,
-}: ILinkDefaultProps) => {
-
-  const points = config.smartRouting ?
-    !!toPort && !!matrix ? generateSmartPath(matrix, startPos, endPos, fromPort, toPort) : generateRightAnglePath(startPos, endPos)
+export const LinkDefault = (props: ILinkDefaultProps) => {
+  const { config, startPos, endPos, fromPort, toPort, matrix } = props
+  const points = config.smartRouting
+    ? !!toPort && !!matrix
+      ? generateSmartPath(matrix, startPos, endPos, fromPort, toPort)
+      : generateRightAnglePath(startPos, endPos)
     : generateCurvePath(startPos, endPos)
 
-  const linkColor: string = (fromPort.properties && fromPort.properties.linkColor) || 'cornflowerblue'
+  const linkColor: string =
+    (fromPort.properties && fromPort.properties.linkColor) || 'cornflowerblue'
 
-  return (
-    <svg
-      className={toPort ? 'react-flow-chart__link--placed' : 'react-flow-chart__link--dragging'}
-      style={{ overflow: 'visible', position: 'absolute', cursor: 'pointer', left: 0, right: 0 }}
-    >
-      <circle
-        r="4"
-        cx={startPos.x}
-        cy={startPos.y}
-        fill={linkColor}
-      />
-      {/* Main line */}
-      <path
-        d={points}
-        stroke={linkColor}
-        strokeWidth="3"
-        fill="none"
-      />
-      {/* Thick line to make selection easier */}
-      <path
-        d={points}
-        stroke={linkColor}
-        strokeWidth="20"
-        fill="none"
-        strokeLinecap="round"
-        strokeOpacity={(isHovered || isSelected) ? 0.1 : 0}
-        onMouseEnter={() => onLinkMouseEnter({ config, linkId: link.id })}
-        onMouseLeave={() => onLinkMouseLeave({ config, linkId: link.id })}
-        onClick={(e) => {
-          onLinkClick({ config, linkId: link.id })
-          e.stopPropagation()
-        } }
-      />
-      <circle
-        r="4"
-        cx={endPos.x}
-        cy={endPos.y}
-        fill={linkColor}
-      />
-    </svg>
-  )
+  const className = toPort ? 'react-flow-chart__link--placed' : 'react-flow-chart__link--dragging'
+
+  const linkProps = {
+    config,
+    points,
+    linkColor,
+    startPos,
+    endPos,
+    className,
+    ...props,
+  }
+  
+  return config.showArrowHead
+    ? <ArrowLink {...linkProps} />
+    : <RegularLink {...linkProps} />
 }