+++ src/components/Port/Port.wrapper.tsx
@@ -1,8 +1,7 @@
 import { isEqual } from 'lodash'
 import * as React from 'react'
-import * as ReactDOM from 'react-dom'
 import { v4 } from 'uuid'
-import { IConfig, ILink, INode, IOnLinkCancel, IOnLinkComplete, IOnLinkMove, IOnLinkStart, IOnPortPositionChange, IPort, IPosition, ISelectedOrHovered } from '../../'
+import { IConfig, ILink, INode, IOnLinkCancel, IOnLinkComplete, IOnLinkMove, IOnLinkStart, IOnPortPositionChange, IPort, IPosition, ISelectedOrHovered, ISize } from '../../'
 import CanvasContext from '../Canvas/CanvasContext'
 import { IPortDefaultProps, PortDefault } from './Port.default'
 
@@ -26,6 +25,7 @@ export interface IPortWrapperProps {
   hoveredLink: ILink | undefined
   port: IPort
   node: INode
+  portsSize: ISize
   onPortPositionChange: IOnPortPositionChange
   Component: React.FunctionComponent<IPortDefaultProps>
 
@@ -49,7 +49,9 @@ export class PortWrapper extends React.Component<IPortWrapperProps> {
   public componentDidUpdate (prevProps: IPortWrapperProps) {
     // Update port position after a re-render if there are more ports on the same side
     // or if node.size has changed
-    if (this.portsOfType(this.props) !== this.portsOfType(prevProps) || !isEqual(this.props.node.size, prevProps.node.size)) {
+    if (this.portsOfType(this.props) !== this.portsOfType(prevProps)
+        || !isEqual(this.props.node.size, prevProps.node.size)
+        || !isEqual(this.props.portsSize, prevProps.portsSize)) {
       this.updatePortPosition()
     }
   }
@@ -162,7 +164,7 @@ export class PortWrapper extends React.Component<IPortWrapperProps> {
   }
 
   private updatePortPosition () {
-    const el = ReactDOM.findDOMNode(this.nodeRef.current) as HTMLInputElement
+    const el = this.nodeRef.current as HTMLInputElement
     if (el) {
       // Ports component should be positions absolute
       // Factor this in so we get position relative to the node