+++ src/types/chart.ts
@@ -1,59 +1,76 @@
 import { IPosition, ISize } from './generics'
 
-export interface IChart {
-  offset: IPosition
-  nodes: {
-    [id: string]: INode,
-  }
-  links: {
-    [id: string]: ILink,
-  }
-  scale: number
-  properties?: any
-
-  /** System Temp */
-  selected: ISelectedOrHovered
-  hovered: ISelectedOrHovered
-}
+export type IChart<
+  ChartProps = undefined,
+  NodeProps = undefined,
+  LinkProps = undefined,
+  PortProps = undefined
+  > = {
+    offset: IPosition
+    nodes: {
+      [id: string]: INode<NodeProps, PortProps>;
+    }
+    links: {
+      [id: string]: ILink<LinkProps>;
+    }
+    scale: number
+    /** System Temp */
+    selected: ISelectedOrHovered
+    hovered: ISelectedOrHovered,
+  } & (ChartProps extends undefined ? {
+    properties?: any,
+  } : {
+    properties: ChartProps,
+  })
 
 export interface ISelectedOrHovered {
   type?: 'link' | 'node' | 'port'
   id?: string
 }
 
-export interface INode {
+export type INode<NodeProps = undefined, PortProps = undefined> = {
   id: string
   type: string
   position: IPosition
   orientation?: number
+  readonly?: boolean
   ports: {
-    [id: string]: IPort,
+    [id: string]: IPort<PortProps>;
   }
-  properties?: any
   /** System Temp */
-  size?: ISize
-}
+  size?: ISize,
+} & (NodeProps extends undefined ? {
+  properties?: any,
+} : {
+  properties: NodeProps,
+})
 
-export interface IPort {
+export type IPort<PortProps = undefined> = {
   id: string
   type: string
   value?: string
-  properties?: any
   /** System Temp */
-  position?: IPosition
-}
+  position?: IPosition,
+} & (PortProps extends undefined ? {
+  properties?: any,
+} : {
+  properties: PortProps,
+})
 
-export interface ILink {
+export type ILink<LinkProps = undefined> = {
   id: string
   from: {
     nodeId: string
     portId: string,
   }
   to: {
-    nodeId?: string
-    portId?: string
+    nodeId?: string;
+    portId?: string;
     /** System Temp */
-    position?: IPosition,
-  }
-  properties?: any
-}
+    position?: IPosition;
+  },
+} & (LinkProps extends undefined ? {
+  properties?: any,
+} : {
+  properties: LinkProps,
+})