+++ stories/components/index.ts
@@ -1,4 +1,6 @@
+export * from './Code'
 export * from './Content'
+export * from './Message'
 export * from './Page'
 export * from './Sidebar'
 export * from './SidebarItem'