+++ stories/index.tsx
@@ -1,9 +1,9 @@
 import { storiesOf } from '@storybook/react'
 import * as React from 'react'
-
 import { ConfigSnapToGridDemo } from './ConfigSnapToGrid'
 import { ConfigValidateLinkDemo } from './ConfigValidateLink'
 import { CustomCanvasOuterDemo } from './CustomCanvasOuter'
+import { CustomGraphTypes } from './CustomGraphTypes'
 import { CustomLinkDemo } from './CustomLink'
 import { CustomNodeDemo } from './CustomNode'
 import { CustomNodeInnerDemo } from './CustomNodeInner'
@@ -12,7 +12,10 @@ import { DragAndDropSidebar } from './DragAndDropSidebar'
 import { ExternalReactState } from './ExternalReactState'
 import { InternalReactState } from './InternalReactState'
 import { LinkColors } from './LinkColors'
+import { NodeReadonly } from './NodeReadonly'
+import { LinkWithArrowHead } from './LinkWithArrowHead'
 import { ReadonlyMode } from './ReadonlyMode'
+import { SelectableMode } from './SelectableMode'
 import { SelectedSidebar } from './SelectedSidebar'
 import { SmartRouting } from './SmartRouting'
 import { StressTestDemo } from './StressTest'
@@ -40,5 +43,9 @@ storiesOf('Other Config', module)
   .add('Snap To Grid', ConfigSnapToGridDemo)
   .add('Link validation function', ConfigValidateLinkDemo)
   .add('Read only mode', ReadonlyMode)
+  .add('Node read only', NodeReadonly)
+  .add('Selectable Mode', SelectableMode)
   .add('Smart link routing', SmartRouting)
   .add('Zoom', () => <Zoom />)
+  .add('Type-safe properties', CustomGraphTypes)
+  .add('Link arrow heads',() => <LinkWithArrowHead />)